Clean up JdbcSqlListExecutionRepository.
