+++ common/src/main/java/org/opentestsystem/rdw/ingest/common/repository/JdbcSqlListExecutionRepository.java
@@ -1,0 +1,38 @@
+package org.opentestsystem.rdw.ingest.common.repository;
+
+import org.springframework.jdbc.core.JdbcOperations;
+import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
+import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
+import org.springframework.jdbc.core.namedparam.SqlParameterSource;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.Map;
+
+/**
+ * JDBC implementation of a SqlListExecutionRepository.
+ */
+public class JdbcSqlListExecutionRepository implements SqlListExecutionRepository {
+    private final NamedParameterJdbcTemplate jdbcTemplate;
+
+    public JdbcSqlListExecutionRepository(final NamedParameterJdbcTemplate jdbcTemplate) {
+        this.jdbcTemplate = jdbcTemplate;
+    }
+
+    @Transactional
+    @Override
+    public void execute(final Iterable<String> sqls) {
+        final JdbcOperations operations = jdbcTemplate.getJdbcOperations();
+        for (final String sql : sqls) {
+            operations.execute(sql);
+        }
+    }
+
+    @Transactional
+    @Override
+    public void execute(final Iterable<String> sqls, final Map<String, ?> parameters) {
+        final SqlParameterSource params = new MapSqlParameterSource(parameters);
+        for (final String sql : sqls) {
+            jdbcTemplate.update(sql, params);
+        }
+    }
+}
