+++ group-processor/src/main/java/org/opentestsystem/rdw/ingest/group/repository/JdbcSqlListExecutionRepository.java
@@ -1,37 +1,17 @@
 package org.opentestsystem.rdw.ingest.group.repository;
 
-import org.opentestsystem.rdw.ingest.common.repository.SqlListExecutionRepository;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.jdbc.core.JdbcOperations;
-import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
-import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.stereotype.Repository;
-
-import java.util.Map;
 
 /**
  * JDBC implementation of a SqlListExecutionRepository.
  */
 @Repository
-public class JdbcSqlListExecutionRepository implements SqlListExecutionRepository {
+public class JdbcSqlListExecutionRepository extends org.opentestsystem.rdw.ingest.common.repository.JdbcSqlListExecutionRepository {
 
     @Autowired
-    private NamedParameterJdbcTemplate jdbcTemplate;
-
-    @Override
-    public void execute(final Iterable<String> sqls) {
-        final JdbcOperations operations = jdbcTemplate.getJdbcOperations();
-        for (final String sql : sqls) {
-            operations.execute(sql);
-        }
-    }
-
-    @Override
-    public void execute(final Iterable<String> sqls, final Map<String, ?> parameters) {
-        final SqlParameterSource params = new MapSqlParameterSource(parameters);
-        for (final String sql : sqls) {
-            jdbcTemplate.update(sql, params);
-        }
+    public JdbcSqlListExecutionRepository(final NamedParameterJdbcTemplate jdbcTemplate) {
+        super(jdbcTemplate);
     }
 }
