+++ migrate-reporting/src/main/java/org/opentestsystem/rdw/ingest/migrate/reporting/repository/impl/JdbcReportingSqlListExecutionRepository.java
@@ -1,42 +1,18 @@
 package org.opentestsystem.rdw.ingest.migrate.reporting.repository.impl;
 
+import org.opentestsystem.rdw.ingest.common.repository.JdbcSqlListExecutionRepository;
 import org.opentestsystem.rdw.ingest.common.repository.SqlListExecutionRepository;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.stereotype.Repository;
-import org.springframework.transaction.annotation.Transactional;
-
-import java.util.Map;
 
 /**
  * Jdbc implementation of {@link SqlListExecutionRepository} on the reporting datamart
  */
 @Repository
-class JdbcReportingSqlListExecutionRepository implements SqlListExecutionRepository {
+class JdbcReportingSqlListExecutionRepository extends JdbcSqlListExecutionRepository {
 
-    @Autowired
-    @Qualifier("reportingJdbcTemplate")
-    private NamedParameterJdbcTemplate jdbcTemplate;
-
-    @Transactional
-    @Override
-    public void execute(final Iterable<String> sqls) {
-
-        for (final String sql : sqls) {
-            jdbcTemplate
-                    .getJdbcOperations()
-                    .execute(sql);
-        }
-    }
-
-    @Transactional
-    @Override
-    public void execute(final Iterable<String> sqls, final Map<String, ?> parameters) {
-
-        for (final String sql : sqls) {
-            jdbcTemplate
-                    .update(sql, parameters);
-        }
+    public JdbcReportingSqlListExecutionRepository(@Qualifier("reportingJdbcTemplate") final NamedParameterJdbcTemplate jdbcTemplate) {
+        super(jdbcTemplate);
     }
 }
